RDKBDEV-925: drop support for obsolete _COSA_DRG_CNS_
Reason for change:
drop support for obsolete _COSA_DRG_CNS_ and _COSA_DRG_TPG_ platform
These are legacy Cisco platforms which are not supported or buildable
in RDKB. They predate the introduction of the RDKB HAL APIs so the
code isn't generally useful to keep around, even as a reference.

  for file in $(find . -type f -name '*.[ch]') ; do unifdef -U_COSA_DRG_TPG_ -U_COSA_DRG_CNS_ -o $file $file ; done

Test Procedure: Build and basic sanity
Risks: None.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>
Change-Id: I0afcc59b7577b4cde37e93c6bb2db18f75951872